+++ NEWS.md
@@ -1,4 +1,4 @@
-# soilDB 2.6.4 (2021-07-23)
+# soilDB 2.6.4 (2021-08-06)
  * `fetchNASIS(from="pedons")` now supports `fill=TRUE` argument just like `from="components"` to include pedons that have no horizon records
  * `createStaticNASIS()`: column order should match NASIS, even if data types require reorder for ODBC driver
  * `fetchSoilGrids()` bug fixes, updates to metadata and references in documentation (https://github.com/ncss-tech/soilDB/issues/201)
@@ -7,7 +7,8 @@
    * Improved error handling
    * Added `progress` and `verbose` arguments for text progress bar and additional message output
    * Added support for {sf} and {sp} POINT geometry inputs
-   
+ * Add `get_SDA_coecoclass` SOD-style method for mapunit/component level summaries of ecological site and other vegetation class information
+ 
 # soilDB 2.6.3 (2021-07-22)
  * `SDA_query()` and all functions that call `SDA_query()` get proper column class handling (related to #190), however:
    - be careful with the use of CAST(): unknown datatypes may not be correctly interpreted